Enabled connetion for dfu and added commands for fetch records, Renamed file names
