+++ move/lib/globals.dart
@@ -54,6 +54,8 @@ class hPi4Global {
   static const List<int> TempTrend = [0x03];
   static const List<int> ActivityTrend = [0x04];
 
+  static const List<int> ECGRecord = [0x10];
+
   static const int CES_CMDIF_TYPE_LOG_IDX = 0x05;
   static const int CES_CMDIF_TYPE_DATA = 0x02;
   static const int CES_CMDIF_TYPE_CMD_RSP = 0x06;
